fix(plugins) Fix URL used to connect to social auth
Social Auth based plugins need to base their URLs on the root domain and
not a customer domain. This ensures that existing Oauth client
configuration continues to work. This is relevant for the asana plugin.

I've also updated the UI components used in plugin warnings as they were
using bootstrap based styles.

Refs HC-663