fix(codecov): Fix git blame for multi repo projects
In some projects where there are more than one repository associated to a project, we get a commit that comes from the release.
For instance, in the case of the sentry project, we get the commit from the getsentry repo, thus, we can't use that commit with the sentry repo.

Fixes:

* Even if a commit is set, run the git blame experiment
  * As mentioned above, the commit may not even be for the right repo
* Report an error if we fail to git blame