test(js): Provide a default legacy context in RTL render (#44801)
There are unfortunately multiple consumers of legacy context still in
the app, such as `withOrganization` and react-router. Tests are
sometimes broken without passing in `context` to the RTL `render`
function because of this, which can be incredibly confusing. For
example, `Link`s render without an href.

This change simply provides a default so tests are simpler to write.