fix(replay): Fix duration due to incorrect LCP endtimestamp (#44842)
Due to 7.38.0 (https://github.com/getsentry/sentry-javascript/pull/7187)
LCP endtimestamp was being incorrectly recorded. The end timestamp is
used in our UI to calculate duration at the moment, so LCP with high
values are causing our calculated duration to be incorrect. Ignore LCP
in the duration calculation as a temporary fix. Ideally we would get SDK
timestamps correct so that we only have the event duration as the source
of truth for duration.