fix(indexer): Update join method of SimpleProduceStep
Arroyo 2.5.1 introduced a fix to Reduce causing this method to actually
be called. join() was never being called from the previous step currently.

When we initially deployed Arroyo 2.5.1 we noticed LeaveGroupRequestTimeouts
as the timeout exceeded 5000 ms. (This caused INC-311 at Sentry).

This change logs the time the produce step takes to close. It also reduces
the timeout to try and prevent INC-311 happening again.

We should merge it before trying Arroyo 2.5.1 again.