chore(derived_code_mappings): Add extra logging (#44825)
A couple of weeks ago, I uncovered (or caused) [an issue](https://sentry.sentry.io/issues/3915004865) that we were not aware of. This adds some logging to help understand why we're getting an empty list.

The call to `get_repositories` from Github is coming as an empty list of repositories, thus, we can't do any processing for those customers.