ref: Increase reprocessing TTL (#44883)
This increases (and refreshes) the main reprocessing counter/info in
Redis. It also introduces a separate TTL for tombstones.

Reprocessing has been getting "stuck" quite a lot recently. Part of the
problem is this counter which is responsible for calling
`finish_reprocessing` in the end, and a missing counter value will also
manifest itself as a progress bar stuck at 100% which is confusing to
customers.

We have seen reprocessing jobs which are unreasonably slow, and can
legitimately take >24h to complete. Increasing and refreshing the TTL
reduces the likelihood of it breaking reprocessing, and also allows the
team more time to investigate potential issues with reprocessing.