fix: Put project transfer links back on root domain
We can't use customer-domains for project transfers as the receiver may
not be in the receiving org, and the reciever may have multiple
organizations so we can't pre-select one either.

Longer term this will need to become a region based URL but we don't
have those in use yet.

Fixes #44823