fix(view-hierarchy): Clicking on the right of overflown row labels not selecting (#44907)
The rows don't extend the full width if they require scrolling to the
right. This means the onClick handler doesn't trigger when clicking in
the whitespace after the row labels end for elements that need to be
scrolled horizontally into view. The hook still calculates a click in
the whitespace as a selection so we can hook into this for a workaround
to set the selected node when we detect a change.

Closes #44754