feat(domains) Add test to ensure that future routes work with customer-domains (#44922)
Going forward we need to define both org slug based routes and routes
for customer domains. Currently that requirement is being held up with
code review. This change adds some automation to ensure that for each
path based route we also have a domain based route defined.

I figured having checks for the routes being defined is enough but I can
go further and attempt to resolve the components and verify that those
are the same too if folks feel that level of check is relevant.

Refs HC-668

---------

Co-authored-by: Alberto Leal <mail4alberto@gmail.com>