feat(perf-issues): respect span render blocking status in detector
As of browser SDK 7.38.0, we now log the render-blocking status reported by the
browser in resource spans' `data` hash. If a value is present and that value is
`non-blocking`, ignore our heuristics and treat the value as non-blocking.

This commit also updates each test case to start from a valid event and then
minimally mutate it into an invalid one, because I was losing confidence that
each test was really testing what it should.