fix(issue-platform): Make sure generic events have a title (#44974)
This uses the `issue_title` from the occurrence as the default title on
the event. This allows events in discover to display a useful title.

I don't think we have a case where we'll want to allow users to pass an
event title that's different to the occurrence title, but we can add
that if someone wants it later.

Note that for perf issues, they'll be using existing transactions and so
these events will just be transactions and have their original title.