feat(perf-issues): respect span render blocking status in detector (#44981)
As of browser SDK 7.38.0, we now [log the render-blocking
status](https://github.com/getsentry/sentry-javascript/pull/7127)
reported by the browser in resource spans' `data` hash. (Background
info:
[MDN](https://developer.mozilla.org/en-US/docs/Web/API/PerformanceResourceTiming/renderBlockingStatus)).
If a value is present and that span is `non-blocking`, ignore our
heuristics and treat the resource as non-blocking in the Large Render
Blocking Asset detector.

This commit also updates each test case to start from a valid event and then
minimally mutate it into an invalid one, because I was losing confidence
that each test was really testing what it should.