feat: Add the `kafka-options` decorator
This is intended to replace the `batching_kafka_options` decorator
for consumers that don't actually do batching. Eventually, `batching_kafka_options`
should be updated to only include those options actually related to batching
(i.e. --max-batch-size and --max-batch-time).

This PR also updates the indexer to remove some of the options being passed.
There were previously way too many types of batching options being passed in here,
and it was very unexpected that --max-batch-size and --max-batch-time had no impact
on the actual batch size but only controlled the commit policy. Configuring the commit
policy separately to the batch size should not be necessary.