chore(hybrid-cloud): Break ProjectIntegration fk (#44992)
Accomplishes two goals:

1. Setups Integration.delete to properly cascade through
HybridCloudForeignKey @leeandher
2.  Uses outboxes to perform the syncing @dashed 
3. Incidentally breaks the ProjectIntegration.integration foreign key,
which needs to be broken anyways.

It is understood that ProjectIntegration isn't currently being used and
is a deprecated model, but the reality is that it is a model and table
that exists, and thus its foreign key must be broken as part of our
hybrid cloud work. It also is a good target for 1 and 2 above because
breaking it can't have any production impact otherwise.