chore(profiling): Reduce date range to 24h on profiling (#45132)
Similar motivations to why performance initially used 24h as the default
date range. Since we're using the transactions dataset for profiling
now, querying 14d of data is a lot and can be slow, so reduce it to 24h
by default. Users can still choose to select 14d if they explicitly
change the date filters.