ref(deletions): Delete metric alerts when project is deleted (#45086)
When a project is deleted, we don't delete the project's metric alert
rules. The table is actually set up to allow multiple projects to map to
a metric alert rule, but the feature was never implemented that way.
Regardless, I'm checking the `include_all_projects` boolean to future
proof it in case we ever allow that. I'll follow this up with a
migration to remove all the old metric alert rules that belong to
deleted projects.