feat(codecov): Enable PR comment and disable patch check
Unfortunately, the backend patch check triggers early. As soon as there are five jobs that upload coverage data.
If you are watching at your PR while the CI is running, you will get a red check is likely not to be updated with the final results.

This change makes the check always be green, thus, not distracting engineers while also adding a comment to the PR with coverage information.