feat(hc): Pull RPC service delegation into abstract base class
For hybrid cloud ervice classes that will go between silos as remote
procedure calls, introduce a new RpcService base class and replace the
`silo_mode_delegation` function with declarations in the immediate
subclass.

Implement stub methods in RpcService for making remote calls. The motive
for moving the "local mode" declaration into the class is, in part, to
make it visible to these class methods, so that it can be used for
resolving to the correct region.