chore(issue assignment): Add logging for`GroupOwner` auto assignment (#45142)
There's currently a bug in our issue assignment logic such that suspect commits are sometimes passed over in favor of issue ownership rules or codeowners when auto-assigning an issue.

For folks with Commit Context enabled, the calculating of suspect commits involves an API call to an outside service (GitHub or GitLab), meaning it's possible that this problem is caused by a race condition: when assignment happens, the data for suspect commit calculation may or may not have come back from the API yet.

To test this hypothesis, it would be helpful to know when each stage of the process happened for an event with the wrong assignee. We already log the completion of suspect commit processing, but we don't log the completion of either ownership rule/code owners processing or the completion of auto-assignment. This PR adds that logging.

Once it's merged, we'll need to look at logs for `process_commit_context.success` , `handle_owner_assignment.success`, and `handle_auto_assignment.success` .