feat(codecov): Enable PR comment and make patch check always green (#45186)
Unfortunately, the backend patch check triggers early. As soon as there are five jobs (current after_n_builds set) that upload coverage data the check will be triggered. We use a value of 5 because that's the number of FE checks that upload coverage data.

If you are watching the CI running for your PR, you will get at first a red check with bad coverage. It will get updated later.

This change makes the check alwaysgreen, thus, not distracting engineers while also adding a comment to the PR with coverage information.

Once we have support for after_n_build for flags we can re-enable it.