fix(perf): Only show event "Status" meta on backend projects (#45096)
This is a continual source of confusion, most recently in #43428. We've
been doing some hoop-jumping with this, but the "Status" meta on the
event details page only applies to backend transactions, so we're only
going to show it for backend transactions.

This mostly undoes #34007 with a twist: hides the meta value from
non-backend events.