feat(monitors): Add basic ingestion via Kafka (#45079)
Adds an ingest consumer similar to Replays and Profiles that reads from
a Kafka topic `ingest-monitors` and updates them directly in postgres.
The consumer runs automatically in devserver and can be started via
`sentry run ingest-monitors`.

This supports creating and updating check ins with a status and
duration. Creating monitors is not supported yet. The consumer closely
follows the logic from checkin creation, but there is a difference for
disabled monitors compared to the Update (PUT) endpoint.