chore(profiling): Remove x axis preference from flamegraph
We should always align to the transaction's x axis when it exists and fall back
to the profile's x axis when it doesn't exist. In practice, the transaction
should always exist. So we're removing the option to switch between the the 2
options as it can leave the flamegraph in a strange state.