bug(replays): Use the new `/replays-count/?returnIds=true` query to list replays related to issues (#45271)
This follows the backend work done in
https://github.com/getsentry/sentry/pull/44774

Basically we're asking now for a list of replays, that are related to
this issue, and where the replays are guaranteed to exist.

Before we were asking for issues with replay_id attached, but the replay
itself might be missing (there are many reasons it could be gone. ie:
deleted).

For this example issue it was common to get back a list of 3 replays,
even though the title suggested that there were 13.
After we see all 13:
<img width="1277" alt="SCR-20230301-j4r"
src="https://user-images.githubusercontent.com/187460/222273029-aae790d0-b04b-4309-942a-6d1f38ebc0c8.png">

I also checked against an example issue that has over 50 replays in the
count, we do still see 50 in the table. As before there is no
pagination, so 50 is the maximum that people will see.

50 replays in the table means 50 examples of the issue. It should be
enough to start a debugging session.

Fixes #44267