ref(hc): Remove future annotations in HC service declarations (#45281)
The motive is to be able to reflect on type annotations when
automatically wiring the services into remote procedure calls, including
serializing and deserializing arguments. The main benefit is to fail
loudly in case a type annotation makes a forward reference.

This would ideally be a temporary convention. The problem would go away
if we upgrade to Python 3.10.