ref(codeowners): remove owners/api from unowned endpoints (#45320)
The original intention of this codeowners grouping was to review unowned
code API / serializer code. unfortunately, there is _a lot_ of this, and
I find this floods my inbox, and do not have enough time to review all
of this code. owners-api will still be tagged on new endpoints and
tooling for APIs. folks who would like a generic reviewer on an API can
still tag the group, or `app-backend`.