fix(domains) Improve devserver workflow with customer-domains
When you have customer-domains active on a sentry-only dev environment,
you currently get stuck in an infinite redirect loop as the application
attempts to send you to the organization domain, but fails to extract
the subdomain as it lacks the base_hostname configuration to discover
what the subdomain is.

Defaulting these values to 'localhost' is potentially contentious as
I know that some staff prefer to use `dev.getsentry.net` as their base
domain. These defaults are also incompatible with the `--bind` option to
`sentry devserver`.

These changes shouldn't impact getsentry dev as it defines these values,
nor should it impact single-tenant and self-hosted as they don't run
within devserver.