feat(custom_instrumentation): Set possible_mistag and start `bind_organization_context` span
Using possible_mistag is useful to track which transactions are setting an org slug tag when it's already set to something different (read: leak of tags).

Starting a span for `bind_organization_context` allows seeing it more easily in transactions and perhaps helping us find the root of where we leak.