bug(perf-issues): hook up N+1 Query detection rate again (#45380)
The `n_plus_one_db_detection_rate` project option was being used to
disable N+1 detection on projects with fingerprint issues. At some point
the N+1 Query detector stopped respecting that option. Hook it back up,
and hook up the MN+1 Query detector to it as well (as they create the
same type of issue). If we have a reason in the future to split that
option into two, we might, but for now they seem to share fingerprinting
issues (typically dynamic custom parent span descriptions).