ref(profiling): Add optional request header for profiling
This is the opt-in introduction of the document-policy header required to enable browser profiling.

Since js-self-profiling has some overhead this is added as a conditional response header only if the request header is present for the time being, which should allow some of us to dogfood browser profiling from the upcoming sdk browser profiling package.

Required request header to enable this will be 'x-sentry-browser-profiling' although I'm open to other suggestions.