ref: Update MockApiClient to have an asyncDelay flag, so test run with timers (#45330)
This is a bit of a followup to
https://github.com/getsentry/sentry/pull/45228. In that PR i noticed
that we can (woo!) test mocked api responses that are still pending if
we first set `MockApiClient.mockAsync = true;`

That is cool and helpful. But it's limiting because all queued api calls
will be resolved after 1ms, so if we use `jest.runAllTimers()` and/or
`jest.advanceTimersByTime` it's the same result: we can't control
exactly which api responses are resolved, it's everything that's queued
or nothing.
It should be possible to control when some responses complete, so that
we can test parallel requests that take different amounts of time.

This update makes that possible, by updating the type of `mockAsync`
(and the name, it's `asyncDelay` now) we can set the number of ms that
an api response should be held back.

So now we've got a new `AddMockResponseOptions` type that can be passed
when calling `MockApiClient.addMockResponse()`. Barring you can still
set `MockApiClient.asyncDelay = number` which will act as a fallback. In
either case the current value is cached/closed over so when it's time to
match the api request with the response, the response will behave as
instructed. I'm not sure about whether `wrapCallback` can be called
directly, or from inside a text, but i'll close over the current
`MockApiClient.asyncDelay` singleton value.

Also i expanded the `useReplayData` test case so it shows the
progressive data loading that we expect when one request is finished,
but the next request is not. -> errors