ref(relay): Stop sending unnecessary fields in project config (#45124)
Make sure that the project configs that Sentry generates are consistent
with what Relay serializes for downstream usage:

* Sampling rules in Relay do not have an `active` field, so stop sending
it.
* `glob` sampling rules do not have options, so stop sending them.
* Sentry unnecessarily sends some empty/default project config fields
that Relay can easily derive. Stop sending them.

**Why?**

* https://github.com/getsentry/relay/pull/1860 introduced a validation
function that can be used Sentry-side to check if Sentry accidentally
produces any fields that Relay cannot use. This has already surfaced
some inconsistencies, but it requires that Sentry and Relay produce the
same fields.
* Omitting default values reduces the payload size both in redis and on
the wire.