feat(modals): Add option to round corner of highlight corners modal container (#45506)
this pr adds rounding the corners of the highlightModalContainer
illustrations . Right now, if you use highlightModalContainer, then you
can set `overflow: hidden` to round the corners of the modal, but if you
use the close modal button then you can't have `overflow: hidden` or it
will hide part of the button. Getting rid of `overflow: hidden` will
expose the bottom left corner of the illustration. this pr will round
the corners so the `overflow` value doesn't impact the illustrations.