feat(issue-platform): Add ability to control how grouptypes are released via features
This adds `is_visible`, `allow_ingest` and `allow_post_process_group` to the `GroupType` class.
These are unused at the moment, but will be used to control how a new issue type is released to
customers.

The way it works is that we now have `released` on each group type. When we're fine to release a
group type everywhere (including self-hosted and ST) then we can set this to True and it'll be
available everywhere. Before we fully release a group type, we instead rely on feature flags to
control how different parts of the group type are released.

These feature flags all default to false, and they won't be set to True in sentry. In getsentry,
we'll add another feature handler that register options for allow_ingest and
allow_post_process_group. That way we can partially roll out to LA/EA/GA customers. `is_visible`
controls whether these issue types are returned from issue search and visible in the UI, and just
relies on a flagr feature flag in getsentry.

While I was here I also refactored the group registry to all be contained in a single class. I
noticed that in test_grouptype.py we were initially patching the various lookup dicts correctly, but
since we've added more it has started to diverge. To avoid weird problems in the future where we
pollute between tests I consolidated into a single registry that can be patched. It also keeps all
these dicts in one place, and keeps the logic to populate them contained. Might make a refactoring
pr at some point to remove the old shims and use the registry directly.