fix(monitors): Allow checkin updates using slugs
This was broken and missing a test case.

The issue here is that the MonitorCheckInEndpoint base class was not
passing the organization_slug to the parent MonitorEndpoint base class,
thus it was ONLY accepting GUID lookups for monitors.

This was easily fixed by allowing the organization_slug to passthrough
the **kwargs in the super().convert_args(...) call