feat(txnames): 2x threshold, 2x sample size (#45573)
We currently merge a node in the URL tree if it has > 100 children,
assuming they are high-cardinality identifiers. But there are legitimate
cases where websites have > 100 static pages, which we should not treat
as such.

This PR sets the threshold to 200 behind a feature flag.

Increasing threshold requires also increasing the sample size, to keep
the probability of discovering a new rule more or less the same.