ref(login): Translate session expired warning message (#45411)
The goal of this PR is to fix a `TODO` comment in
`sentry/src/constants.py` asking to translate the warning message for
session expiration.

### Legal Boilerplate

Look, I get it. The entity doing business as "Sentry" was incorporated
in the State of Delaware in 2015 as Functional Software, Inc. and is
gonna need some rights from me in order to utilize my contributions in
this here PR. So here's the deal: I retain all rights, title and
interest in and to my contributions, and by keeping this boilerplate
intact I confirm that Sentry can use, modify, copy, and redistribute my
contributions, under Sentry's choice of terms.

Co-authored-by: Chad Whitacre <chadwhitacre@sentry.io>