ref(jira): Remove unused/wontfix things (#45584)
Remove the TODOs around matching on display name since we've decided
it's a bad idea due to possible name collision, as well as remove
references to `use_email_scope` in Jira Server since it's an unused code
path (always resolves to `False`) and the related unused code. We used
to try to keep Jira and Jira Server as similar as possible, but since
the code has been split I don't think it matters anymore.