fix(grouping): Always assign group ids to group hashes (#45600)
We cannot fully explain why the hashes were only assigned to group
hashes, but the end result is that the `GroupHash` table creates a lot
of orphaned hashes. Currently only the secondary grouping created hashes
have the group id associated. This should resolve the issue that the
smooth upgrade does not work.