ref(proj-config): Wrap scheduling in a sentry span
Updating a lot of organizations or projects in a single database
transaction causes the `on_commit` list to grow considerably and may
cause memory leaks. It may be very hard to identify these cases.

In an attempt to make these simpler, the `on_commit` call is surrounded
by a sentry span to make it easier to identify on Sentry.