feat(replays): Implement metadata-less storage driver for reading and writing segment data (#44659)
Session Replay is a data intensive product. For this reason we have
reached the limits of our existing filestore infrastructure. Write (and
delete) throughput on the Postgres table will peak soon and there are
questions on the utility of the Filestore service for the Replays use
case. This pull drops usage of Postgres and uploads replays directly to
a blob storage provider.

### Driver System

On ingest we check the configuration and make a determination on which
service-driver should be used. On read, we first check if a
ReplayRecordingSegment record exists. If it does then we read from
Filestore. If not then we read from direct storage.

### Data Model

Segment data is stored in two locations (for the time being). The
Postgres table in the form of ReplayRecordingSegment and Clickhouse in
the form of our replays table. Either of these tables can be used to
identify the existence of a blob in storage (caveat: currently the
Clickhouse row is written without knowledge of the success or failure of
the blob insert - changing this will require a larger refactor to
ingest).

### Notes

- Blobs are TTL'd based on their prefix.  `30/`, `60/`, `90/`.
- No code was added to support prefix deletes on `project_id` or
`replay_id`.
- Currently, user deletes do not delete blob data.
- Blob data is gated by the is-archived value in Clickhouse and is not
user visible.
    - Blobs are eventually deleted by TTL.
- We could delete it, though. I'm wondering if there's risk in letting
it TTL for free.
- Delete requests will now 404 if we detect an archived row in
Clickhouse.
    -  Removes use of the ReplayRecordingSegment model entirely.
- The deletion task will still use ReplayRecordingSegment to delete
filestore records.
- The RecordingSegmentSerializer class was removed.
- Replaced by manually constructing the response output from the
metadata object.
- This was necessary because dataclasses could not be used as a drop-in
replacement from ReplayRecordingSegment instances.
- When using the `StorageBlob.get` method responses are cached in Redis
for 1 hour.
    - Blobs are immutable so we could cache forever™.
- 1 hour was chosen as a reasonable default value. Alternatives are
welcome.
- Most behavior was moved from the API endpoints to the `reader` use
case module.
- Some of the database-querying, performance optimizations added by
@JoshFerge were removed to simplify the interface.
- These can certainly be re-added but since we need to drop Postgres
because of scaling concerns it seems to be not necessary anymore for our
use case.

### Metrics
We'll need to update the dashboards to support the following metrics:

- `replays.lib.storage.FilestoreBlob.get`
- `replays.lib.storage.FilestoreBlob.set`
- `replays.lib.storage.StorageBlob.get`
- `replays.lib.storage.StorageBlob.set`

These will help us understand the performance profile of direct storage.
If it is not performant we can still use the Filestore service without
the File models. This would mean we can't use our own replays specific
bucket (without changes to the filestore service).

### Test Suite

The test suite was partially re-written. We now run each test against
each storage back-end. The storage driver is chosen either by
configuration (ingest) or by the (non-)existence of a
ReplayRecordingSegment row (read).

### Settings

A new `SENTRY_REPLAYS_STORAGE_ALLOWLIST` setting was added. The Sentry
org can be added to this to allow us to test in a more-forgiving,
production environment. Alternatively, we can use a feature flag but I
wasn't sure how to do this.

### Options

Three new options were added:

```
"replay.storage.backend"
"replay.storage.options"
"replay.storage.direct-storage-sample-rate"
```

The `replays.storage.direct-storage-sample-rate` should be 0 until we
have validated the driver with the Sentry org.

### Todo

- [ ] Range-based GDPR deletes.
    - Can leverage Clickhouse rows and delete in an async task.
- [ ] User blob deletes.
- [ ] Scripts for deleting projects and replays by prefix.
    - Optional?
- [ ] Emit Clickhouse row on successful blob ingest.
    - Different unit of work?
- Could also emit to a new Clickhouse table instead of linking the
replay-event and replay-recording-event envelopes together.

---------

Co-authored-by: Josh Ferge <josh.ferge@sentry.io>