bug(replays): Return segment data as streaming output (#45643)
With the deployment of the filestore deprecation PR we noticed a
regression in performance on the recording-segment index page. This is
because we were naively waiting for all the files to download before
returning data to the user. This pull will stream the data segment by
segment as we were originally.