chore(domains) Rename mixed case slugs to lowercase
Rename organization slugs that are mixed case to be all lower case. We
need to do this because mixed case slugs are incompatible with customer
domains. In order to reach 100% on saas (and prepare self-hosted) we
need to rename slugs. This change refines the previous attempt and
handles the case where the 'new' slug is already taken.

Mulligan on #43917

This reverts commit b00e76c8a95ceff0412fa129590091a2bb02329b.