fix(monitors): Widen rate limit window
We got a request to increase number of checkins allowed per second. We're waiting for feedback from
the customer, but I think it's likely that they just have a lot of checkins happening in a short
period of time each minute, and so are bumping into the rate limit. Expanding the window will
probably help here.