refs(monitors): Split legacy ingest endpoints out
This

* splits the GET from the POST in the monitor checkin creation /
  checkin listing endpoint into two.

* splits the GET and POST in the checkin attachment upload / download

The reason we are splitting these is because authentication and
authorization is completely different between these two endpoints, so
really we want to have two totally different base endpoints.

I will be following up with an update to the authentication and base
endpoint classes that should make things much more clear and make
authentication more strict.