feat(onboarding): adds an onboarding backend (#44141)
This PR sets up an onboarding backend that can swap in a different model
for `OrganizationOnboardingTask`. The reason we want this is so the
sandbox can use the onboarding tasks on the organization serializer
which have a different model that's unique per user instead of per org
(so each user can do their own onboarding).