feat(perf-issues): Add an option and project option to send occurrences to platform (#45709)
Create a system-wide and per-project options that would control whether
a performance problem is sent to the occurrences platform. Both options
need to be booleans since the options will be checked first in
`save_aggregate_performance` and then in `send_occurrence_to_platform`
and there's no reliable way to ensure that only one function gets run.