ref(migration): Create shared model NotificationAction (#45509)
This PR introduces a new model `NotificationAction` in place of the one
in [this PR](https://github.com/getsentry/getsentry/pull/9583). Since a
lot of the shape is shared among `AlertRuleTriggerAction` an abstract
model was used for the shared fields.

The way they register handlers is different and I wasn't able to pull
that into the abstractio. However, they share the enums which will much
more explicitly suggest adding handlers for both metric alerts, and
these notifications should we extend the list of services in the future.