fix(domains) Improve devserver + customer domains
If you have system.url-prefix defined as a domain that includes
a customer domain you'll get trapped in a redirect loop when going to
`/organizations/new/` as the application will try to remove the customer
domain, by redirecting to system.url-prefix.

By setting a default value for the option we can improve the workflow
for new devserver environments as they'll have this option
pre-configured.