test(replay): Add unit tests for useReplayData hook (#45444)
Adds new fixtures, and use the to test `useReplayData()`
The new fixtures are directly inspired by the helpers inside of
https://github.com/getsentry/sentry/blob/master/src/sentry/replays/testutils.py#L200

The test/assertions inside of `useReplayData` are also streamlined. We
shouldn't assert that the resulting ReplayReader instance is correct,
it's more direct to assert that the arguments for that class are being
correctly fetched and combined. Testing that class directly happens
elsewhere.

https://github.com/getsentry/sentry/pull/45548 is a followup to remove
some less flexible test stubs, and let these new mocks get more use.

Fixes #38852