chore(domains) Rename mixed case slugs to lowercase (#45669)
Rename organization slugs that are mixed case to be all lower case. We
need to do this because mixed case slugs are incompatible with customer
domains. In order to reach 100% on saas (and prepare self-hosted) we
need to rename slugs. This change refines the previous attempt and
handles the case where the 'new' slug is already taken.

Mulligan on https://github.com/getsentry/sentry/pull/43917

This reverts commit
https://github.com/getsentry/sentry/commit/a4236570d419a531079ac0a994f6c4897d118ec7.