fix(monitors): Widen rate limit window (#45670)
We got a request to increase number of checkins allowed per second, but
the issue here is that this customer has a lot of checkins for different
monitors happening at around the same time. This is causing them to bump
into the rate limit. Instead of increasing our base limit, I'll widen
the window so that we can received 40 * 60 (2400) a minute instead of
40/second.

Fixes #45583