feat(tags): filter device.class from tags endpoint (#45795)
Filters out `device.class` from the tags endpoint. This is because we
already specified `device.class` as a selectable `field` with
description in the frontend, so we can filter out `device.class` from
this endpoint to avoid having it appear twice in column selectors in
areas such as discover and dashboards.